make sure the regression model execution results are stored as float64

+++ python-client/giskard/ml_worker/server/ml_worker_service.py
@@ -1,5 +1,6 @@
 import asyncio
 import logging
+import math
 import os
 import platform
 import posixpath
@@ -12,13 +13,11 @@ from pathlib import Path
 
 import google
 import grpc
-import math
 import numpy as np
 import pandas as pd
 import pkg_resources
 import psutil
 import tqdm
-
 from mlflow.store.artifact.artifact_repo import verify_artifact_path
 
 import giskard
@@ -533,10 +532,20 @@ class MLWorkerServiceImpl(MLWorkerServicer):
             if dataset.target and dataset.target in dataset.df.columns:
                 target_serie = dataset.df[dataset.target]
                 diff = preds_serie - target_serie
-                diff_percent = pd.Series(diff / target_serie, name="diffPercent")
-                abs_diff = pd.Series(diff.abs(), name="absDiff")
+                diff_percent = pd.Series(
+                    diff / target_serie,
+                    name="diffPercent",
+                    dtype=np.float64,
+                )
+                abs_diff = pd.Series(
+                    diff.abs(),
+                    name="absDiff",
+                    dtype=np.float64,
+                )
                 abs_diff_percent = pd.Series(
-                    abs_diff / target_serie, name="absDiffPercent"
+                    abs_diff / target_serie,
+                    name="absDiffPercent",
+                    dtype=np.float64,
                 )
                 calculated = pd.concat(
                     [
